wip: enable, error func

+++ 0_Src/AppSw/Tricore/RVC/TorqueVectoring/TorqueVectoring.h
@@ -16,35 +16,35 @@ typedef struct
 {
 	struct
 	{
-        float v_x;  // enable limit of velocity
-        float delta_sta;    // enable limit of steering angle
-        float k_1;  // gain for enable limit of steering angle
+		double v_x;       // enable limit of velocity
+		double delta_sta; // enable limit of steering angle
+		double k_1;       // gain for enable limit of steering angle
 	} enable;
 
-    struct
-    {
-        float k_2;  // (steering angle) - (wheel gear ratio)
-        float i;    // |(front accel) - (center of mass)|
-    } error;
+	struct
+	{
+		double k_2; // (steering angle) - (wheel gear ratio)
+		double i;   // |(front accel) - (center of mass)|
+	} error;
 
-    struct
-    {
-        float k_p;  // proportional gain
-        float k_i;  // integral gain
-        float k_d;  // derivative gain
-        float k_b;  // inverse calculation constant == k_i
-    } pid;
+	struct
+	{
+		double k_p; // proportional gain
+		double k_i; // integral gain
+		double k_d; // derivative gain
+		double k_b; // inverse calculation constant == k_i
+	} pid;
 
-    struct
-    {
-        float max_delta_trq;    // limit of maximum delta torque
-    } limit;
+	struct
+	{
+		double max_delta_trq; // limit of maximum delta torque
+	} limit;
 
-    struct
-    {
-        float t;    // constant of transfer function constant
-    } tv;
-    
+	struct
+	{
+		double t; // constant of transfer function constant
+	} tv;
+
 } TV_t;
 
 /* Functionc Prototypes */
